dom and logic sepated
